requirements.txt
numba
numpy==1.23.0
nuscenes-devkit
opencv-python-headless
pandas
pytorch-lightning==1.6.0
scikit-image
scipy
setuptools==60.2.0
tensorboardX
mmdet==2.28
mmsegmentation==0.30.0
mmdet3d==1.0.0rc4
mmengine==0.10.1
--find-links https://download.openmmlab.com/mmcv/dist/cu116/torch1.13.0/index.html
mmcv-full==1.7.0